ran simulation and did some typing

+++ src/Agent_Based_Traffic_Simulation/core/Lane.py
@@ -3,9 +3,9 @@ import numpy as np
 class Lane:
 
     # All units will be in millimeters
-    def __init__(self, start_position: np.array, end_position:np.array, lane_width: int):
+    def __init__(self, start_position: np.array, end_position:np.array, lane_width: float):
         self.start_position:np.array = start_position
         self.end_position:np.array = end_position
-        self.lane_width:int = lane_width
+        self.lane_width:float = lane_width
